removed logic from widget
